Refactor configuration and logging components; update agent details, enhance log extraction with model ID, and add HTML report generation for evaluation metrics.

+++ eval_metric_store.py
@@ -22,7 +22,7 @@ class EvalMetricStore:
                     agent TEXT NOT NULL,
                     model TEXT NOT NULL,
 
-                    response_time_ms REAL,
+                    response_time_ms INTEGER,
                     cost REAL,
                     quality_score REAL,
